simulator/requirements.txt
httpx==0.25.2
pyyaml==6.0.1
faker==20.1.0
asyncio-throttle==1.0.2
pydantic==2.5.0 